+++ Sources/Sentry/SentryBreadcrumbTracker.m
@@ -153,8 +153,9 @@
                     sanitizeViewControllerName:[NSString stringWithFormat:@"%@", self]];
                 crumb.data = @ { @"screen" : viewControllerName };
 
+                // Adding crumb via the SDK calls SentryBeforeBreadcrumbCallback
+                [SentrySDK addBreadcrumb:crumb];
                 [SentrySDK.currentHub configureScope:^(SentryScope *_Nonnull scope) {
-                    [scope addBreadcrumb:crumb];
                     [scope setExtraValue:viewControllerName forKey:@"__sentry_transaction"];
                 }];
             }