+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## unreleased
 
+- fix: Remove enum specifier for SentryLevel #822
+
+## 6.0.6
+
+- fix: Call beforeBreadcrumb for Breadcrumb Tracker #815
+
+## 6.0.5
+
+- fix: Add eventId to user feedback envelope header #809
+- feat: Manually capturing User Feedback #804
+
+## 6.0.4
 - fix: Sanitize UserInfo of NSError and NSException #770
 - fix: Xcode 12 warnings for Cocoapods #791